% Generated by roxygen2 (4.0.1): do not edit by hand
\name{burn.DPMMclust}
\alias{burn.DPMMclust}
\title{Burning mcmc iterations from a Dirichlet Process Mixture Model}
\usage{
burn.DPMMclust(x, burnin = 0, thin = 1)
}
\arguments{
\item{x}{a \code{DPMMclust} object.}

\item{burnin}{the number of MCMC iterations to burn (default is \code{0}).}

\item{thin}{the spacing at which MCMC iterations are kept.
Default is \code{1}, i.e. no thining.}

\item{gs}{optionnal vector of length \code{n} containing the gold standard
partition of the \code{n} observations to compare to the point estimate.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{DPMMclust} object minus the burnt iterations
}
\description{
Burning mcmc iterations from a Dirichlet Process Mixture Model
}
\author{
Boris Hejblum
}
\seealso{
\link{summary.DPMMclust}
}

