% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cluster_est_Fmeasure}
\alias{cluster_est_Fmeasure}
\title{Gets a point estimate of the partition using the F-measure as the cost function}
\usage{
cluster_est_Fmeasure(c, thin = 1)
}
\arguments{
\item{c}{a list of vector of length \code{n}. \code{c[[j]][i]} is
the cluster allocation of observation \code{i=1...n} at iteration
\code{j=1...N}.}

\item{thin}{integer indicating the thinning of the MCMC output.
Default is \code{1}.}
}
\value{
a \code{list}:
 \itemize{
     \item{\code{c_est}:}{ a vector of length \code{n}. Point estimate of the partition}
     \item{\code{cost}:}{ a vector of length \code{N}. \code{cost[j]} is the cost
     associated to partition \code{c[[j]]}}
     \item{\code{similarity}:}{  matrix of size \code{n x n}. Similarity matrix
     (see \link{similarityMat})}
     \item{\code{opt_ind}:}{ the index of the optimal partition
     among the MCMC iterations.}
 }
}
\description{
Gets a point estimate of the partition using the F-measure as the cost function
}
\author{
Francois Caron, Boris Hejblum
}
\references{
F. Caron, Y.W. Teh, T.B. Murphy. Bayesian nonparametric Plackett-Luce
models for the analysis of preferences for college degree programmes.
To appear in Annals of Applied Statistics, 2014.
http://arxiv.org/abs/1211.5037

D. B. Dahl. Model-Based Clustering for Expression Data via a
Dirichlet Process Mixture Model, in Bayesian Inference for
Gene Expression and Proteomics, K.-A. Do, P. Muller, M. Vannucci
(Eds.), Cambridge University Press, 2006.
}
\seealso{
\link{similarityMat}
}

