% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_est_MBinderN.R
\name{cluster_est_MBinderN}
\alias{cluster_est_MBinderN}
\title{Gets a point estimate of the partition using a modified Binder loss function
The loss function takes in account distance between mixture components using
the  Bhattacharyya distance. Note that components need to be Gaussian.}
\usage{
cluster_est_MBinderN(c, Mu, Sigma, lambda, a = 1, b = 1, logposterior)
}
\arguments{
\item{c}{a list of vector of length \code{n}. \code{c[[j]][i]} is 
the cluster allocation of observation \code{i=1...n} at iteration 
\code{j=1...N}.}

\item{Mu}{is a list of length \code{n} composed of \code{p x l} matrices.
Where \code{l} is the maximum number of components per partition.}

\item{Sigma}{is list of length \code{n} composed of arrays containing a maximum of 
\code{l} \code{p x p} covariance matrices.}

\item{lambda}{is a nonnegative tunning parameter allowing further control over the distance
function.}

\item{logposterior}{vector of logposterior correponding to each 
partition from \code{c} used to break ties when minimizing the cost function}

\item{\code{a}}{and \code{b} are nonnegative constants seen as the unit cost for each
kind of pairwise misclassification.}
}
\value{
a \code{list}: 
 \itemize{
     \item{\code{c_est}:}{ a vector of length \code{n}. Point estimate of the partition}
     \item{\code{cost}:}{ a vector of length \code{N}. \code{cost[j]} is the cost 
     associated to partition \code{c[[j]]}}
     \item{\code{similarity}:}{  matrix of size \code{n x n}. Similarity matrix 
     (see \link{similarityMat})}
     \item{\code{opt_ind}:}{ the index of the optimal partition 
     among the MCMC iterations.}  
 }
}
\description{
Gets a point estimate of the partition using a modified Binder loss function
The loss function takes in account distance between mixture components using
the  Bhattacharyya distance. Note that components need to be Gaussian.
}
\author{
François Caron, Chariff Alkhassim, Boris Hejblum
}
\references{
J.W.Lau and P.J.Green. Bayesian Model-Based Clustering Procedures.
Journal of Computational and Graphical Statistics, Volume 16, Number 3, 
Pages 526-558, 2007.

D.A.Binder. Bayesian cluster analysis. Biometrika Vol. 65, No. 1, pp. 31-38, 1978
}
\seealso{
\link{similarityMat}
}

