% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytoScatter.R
\name{cytoScatter}
\alias{cytoScatter}
\title{Scatterplot of flow cytometry data}
\usage{
cytoScatter(cytomatrix, dims2plot = c(1, 2), scale_log = FALSE,
  xlim = NULL, ylim = NULL, gg.add = list(theme()))
}
\arguments{
\item{cytomatrix}{}

\item{dims2plot}{}

\item{scale_log}{}

\item{xlim}{}

\item{ylim}{}
}
\description{
Scatterplot of flow cytometry data
}

