% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DPM_GibbsSampler_SkewN_test1K.R
\name{DPM_GibbsSampler_SkewN_test1K}
\alias{DPM_GibbsSampler_SkewN_test1K}
\title{Slice Sampling of Dirichlet Process Mixture of skew Normals}
\usage{
DPM_GibbsSampler_SkewN_test1K(z, hyperG0, a, b, N, doPlot = TRUE,
  nbclust_init = 30, plotevery = 1, ...)
}
\arguments{
\item{z}{data matrix \code{d x n} with \code{d} dimensions in rows
and \code{n} observations in columns.}

\item{hyperG0}{prior mixing distribution.}

\item{a}{shape hyperparameter of the Gamma prior
on the parameter of the Dirichlet Process.}

\item{b}{scale hyperparameter of the Gamma prior
on the parameter of the Dirichlet Process.}

\item{N}{number of MCMC iterations.}

\item{doPlot}{logical flag indicating wether to plot MCMC iteration or not.
Default to \code{TRUE}.}

\item{nbclust_init}{number of clusters at initialisation.
Default to 30 (or less if there are less than 30 observations).}
}
\description{
Hyperprior on the concentration parameter of the DP
}
\examples{
rm(list=ls())
library(ggplot2)
#Number of data
n <- 1000
#n <- 2000
set.seed(1234)
#set.seed(123)
#set.seed(4321)


# Sample data
d <- 2
xi <- matrix(nrow=d, ncol=1, c(-1.5, 1))
psi <- matrix(nrow=d, ncol=1, c(3, -2))
p <- 1 # frequence des clusters

sdev <- matrix(nrow=2, ncol=2, c(0.3, 0, 0, 0.3))
c <- rep(0,n)
z <- matrix(0, nrow=d, ncol=n)
for(k in 1:n){
 c[k] = which(rmultinom(n=1, size=1, prob=p)!=0)
 z[,k] <- xi[, c[k]] + psi[, c[k]]*abs(rnorm(1)) + matrix(rnorm(d, mean = 0, sd = 1), nrow=1, ncol=d)\%*\%sdev
 #cat(k, "/", n, " observations simulated\\n", sep="")
}

 # Set parameters of G0
 hyperG0 <- list()
 hyperG0[["b_xi"]] <- rep(0,d) # c(-1.5, 1) for an informative prior instead...
 hyperG0[["b_psi"]] <- rep(0,d) # c(3, -2) for an informative prior instead...
 hyperG0[["kappa"]] <- 0.001
 hyperG0[["D_xi"]] <- 1000 # weak prior
 hyperG0[["D_psi"]] <- 1000 #0.00001 for a strong prior
 hyperG0[["nu"]] <- 0.1
 hyperG0[["lambda"]] <- diag(d)/10 # diag(d)*0.3 for an informative prior instead...

 # hyperprior on the Scale parameter of DPM
 a <- 0.0001
 b <- 0.0001

 # do some plots
 doPlot <- TRUE
 nbclust_init <- 30



 ## Data
 ########
 p <- (ggplot(data.frame("X"=z[1,], "Y"=z[2,]), aes(x=X, y=Y))
       + geom_point()
       + ggtitle("Simple example in 2d data")
       +xlab("D1")
       +ylab("D2")
       +theme_bw())
 p


 ## alpha priors plots
 #####################
 prioralpha <- data.frame("alpha"=rgamma(n=5000, shape=a, scale=1/b),
                         "distribution" =factor(rep("prior",5000),
                         levels=c("prior", "posterior")))
 p <- (ggplot(prioralpha, aes(x=alpha))
       + geom_histogram(aes(y=..density..),
                        colour="black", fill="white")
       + geom_density(alpha=.2, fill="red")
       + ggtitle(paste("Prior distribution on alpha: Gamma(", a,
                 ",", b, ")\\n", sep=""))
      )
 p



 # Gibbs sampler for Dirichlet Process Mixtures
 ##############################################
 MCMCsample_sn <- DPM_GibbsSampler_SkewN_test1K(z, hyperG0, a, b, N=200, doPlot, nbclust_init, plotevery=10)
 MCMCsample_sn$U_xi
 MCMCsample_sn$U_psi
}
\author{
Boris Hejblum
}

