% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPMGibbsSkewT_EB_parallel.R
\name{DPMGibbsSkewT_EB_parallel}
\alias{DPMGibbsSkewT_EB_parallel}
\title{Slice Sampling of Dirichlet Process Mixture of skew  Students's t-distibutions}
\usage{
DPMGibbsSkewT_EB_parallel(Ncpus, type_connec, z, hyperG0, a, b, N, nbclust_init,
  doPlot = TRUE, plotevery = 1, diagVar = TRUE, verbose = FALSE,
  monitorfile = "", ...)
}
\arguments{
\item{z}{data matrix \code{d x n} with \code{d} dimensions in rows 
and \code{n} observations in columns.}

\item{hyperG0}{prior mixing distribution.}

\item{a}{shape hyperparameter of the Gamma prior 
on the parameter of the Dirichlet Process.}

\item{b}{scale hyperparameter of the Gamma prior 
on the parameter of the Dirichlet Process.}

\item{N}{number of MCMC iterations.}

\item{nbclust_init}{number of clusters at initialisation. 
Default to 30 (or less if there are less than 30 observations).}

\item{doPlot}{logical flag indicating wether to plot MCMC iteration or not.
Default to \code{TRUE}.}

\item{verbose}{logical flag indicating wether partition info is 
written in the console at each MCMC iteration.}

\item{monitorfile}{a writable \link{connections} or a character string naming a file to write into, 
to monitor the progress of the analysis.  
Default is \code{""} which is no monitoring.  See Details.}
}
\value{
a object of class \code{DPMclust} with the following attributes: 
 \itemize{
     \item{\code{mcmc_partitions}:}{a list of length \code{N}. Each
      element \code{mcmc_partitions[n]} is a vector of length 
      \code{n} giving the partition of the \code{n} observations.}
     \item{\code{alpha}:}{ a vector of length \code{N}. \code{cost[j]} is the cost 
associated to partition \code{c[[j]]}}
     \item{\code{weights_list}:}{}
     \item{\code{logposterior_list}:}{}
     \item{\code{data}:}{the data matrix \code{d x n} with \code{d} dimensions in rows 
and \code{n} observations in columns.}
     \item{\code{nb_mcmcit}:}{the number of MCMC itertations}
 }
}
\description{
Slice Sampling of Dirichlet Process Mixture of skew  Students's t-distibutions
}
\examples{
rm(list=ls())
library(ggplot2)

#Number of data
n <- 2000
set.seed(123)
#set.seed(4321)


d <- 2
ncl <- 4

# Sample data

sdev <- array(dim=c(d,d,ncl))

xi <- matrix(nrow=d, ncol=ncl, c(-1.5, 1.5, 1.5, 1.5, 2, -2.5, -2.5, -3))
#xi <- matrix(nrow=d, ncol=ncl, c(-0.5, 0, 0.5, 0, 0.5, -1, -1, 1))
psi <- matrix(nrow=d, ncol=4, c(0.4, -0.6, 0.8, 0, 0.3, -0.7, -0.3, -0.8))
nu <- c(100,15,8,5)
p <- c(0.15, 0.05, 0.5, 0.3) # frequence des clusters
sdev[, ,1] <- matrix(nrow=d, ncol=d, c(0.3, 0, 0, 0.3))
sdev[, ,2] <- matrix(nrow=d, ncol=d, c(0.1, 0, 0, 0.3))
sdev[, ,3] <- matrix(nrow=d, ncol=d, c(0.3, 0.15, 0.15, 0.3))
sdev[, ,4] <- .3*diag(2)

 
c <- rep(0,n)
w <- rep(1,n)
z <- matrix(0, nrow=d, ncol=n)
for(k in 1:n){
 c[k] = which(rmultinom(n=1, size=1, prob=p)!=0)
 w[k] <- rgamma(1, shape=nu[c[k]]/2, rate=nu[c[k]]/2)
 z[,k] <- xi[, c[k]] + psi[, c[k]]*rtruncnorm(n=1, a=0, b=Inf, mean=0, sd=1/sqrt(w[k])) + (sdev[, , c[k]]/sqrt(w[k]))\%*\%matrix(rnorm(d, mean = 0, sd = 1), nrow=d, ncol=1)
 cat(k, "/", n, " observations simulated\\n", sep="")
}
 
MCMCsample_st_EB_parallel <- DPMGibbsSkewT_EB_parallel(Ncpus=2, type_connec ="SOCK", z, hyperG0=priorMix, a, b, N=1500, doPlot, plotevery=100, gg.add=list(theme_bw()), verbose=TRUE)
 
 
# Set parameters of EB G0
priorMix <- list()
priorMix[["weights"]] <- p
priorMix[["parameters"]] <- list()  
for(j in 1:ncl){
 priorMix[["parameters"]][[j]] <- list()
 priorMix[["parameters"]][[j]][["b_xi"]] <- xi[,j]
 priorMix[["parameters"]][[j]][["b_psi"]] <- psi[,j]
 priorMix[["parameters"]][[j]][["lambda"]] <- crossprod(sdev[,,j])
 priorMix[["parameters"]][[j]][["kappa"]] <- 0.001
 priorMix[["parameters"]][[j]][["D_xi"]] <- 1
 priorMix[["parameters"]][[j]][["D_psi"]] <- 1
 priorMix[["parameters"]][[j]][["nu"]] <- d + 0.1
}
 
 # hyperprior on the Scale parameter of DPM
 a <- 0.0001
 b <- 0.0001
 
 # do some plots
 doPlot <- TRUE 
 
 
 
 ## Data
 ########
 p <- (ggplot(data.frame("X"=z[1,], "Y"=z[2,]), aes(x=X, y=Y)) 
       + geom_point()
       + ggtitle("Simple example in 2d data")
       +xlab("D1")
       +ylab("D2")
       +theme_bw())
 p
 
 
 # Gibbs sampler for Dirichlet Process Mixtures
 ##############################################
 MCMCsample_st_EB_parallel <- DPMGibbsSkewT_EB_parallel(Ncpus=2, type_connec ="SOCK", z, hyperG0=priorMix, a, b, N=1500, doPlot, plotevery=100, gg.add=list(theme_bw()), verbose=TRUE)
 
 s <- summary(MCMCsample_st_EB, burnin = 500, thin=2)
 print(s)
 plot(s)
 plot_ConvDPM(MCMCsample_st_EB, from=2)

}
\author{
Boris Hejblum
}

