% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DPMGibbsSkewT_test1K}
\alias{DPMGibbsSkewT_test1K}
\title{Slice Sampling of Dirichlet Process Mixture of skew  Students's t-distibutions}
\usage{
DPMGibbsSkewT_test1K(z, hyperG0, a, b, N, doPlot = TRUE, nbclust_init = 30,
  plotevery = 1, diagVar = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{z}{data matrix \code{d x n} with \code{d} dimensions in rows
and \code{n} observations in columns.}

\item{hyperG0}{prior mixing distribution.}

\item{a}{shape hyperparameter of the Gamma prior
on the parameter of the Dirichlet Process.}

\item{b}{scale hyperparameter of the Gamma prior
on the parameter of the Dirichlet Process.}

\item{N}{number of MCMC iterations.}

\item{doPlot}{logical flag indicating wether to plot MCMC iteration or not.
Default to \code{TRUE}.}

\item{nbclust_init}{number of clusters at initialisation.
Default to 30 (or less if there are less than 30 observations).}

\item{verbose}{logical flag indicating wether partition info is
written in the console at each MCMC iteration.}
}
\value{
a object of class \code{DPMclust} with the following attributes:
 \itemize{
     \item{\code{mcmc_partitions}:}{a list of length \code{N}. Each
      element \code{mcmc_partitions[n]} is a vector of length
      \code{n} giving the partition of the \code{n} observations.}
     \item{\code{alpha}:}{ a vector of length \code{N}. \code{cost[j]} is the cost
associated to partition \code{c[[j]]}}
     \item{\code{weights_list}:}{}
     \item{\code{logposterior_list}:}{}
     \item{\code{data}:}{the data matrix \code{d x n} with \code{d} dimensions in rows
and \code{n} observations in columns.}
     \item{\code{nb_mcmcit}:}{the number of MCMC itertations}
 }
}
\description{
Slice Sampling of Dirichlet Process Mixture of skew  Students's t-distibutions
}
\examples{
rm(list=ls())
library(ggplot2)

#Number of data
n <- 2000
set.seed(123)
#set.seed(4321)


d <- 2
ncl <- 4

# Sample data


xi <- matrix(nrow=d, ncol=1, c(-2.5, -3))
psi <- matrix(nrow=d, ncol=1, c(-0.3, -0.7))
nu <- 27
sdev <- .5*diag(2)


c <- rep(1,n)
w <- rep(1,n)
z <- matrix(0, nrow=d, ncol=n)
for(k in 1:n){
 w <- rgamma(n, shape=nu/2, rate=nu/2)
 z[,k] <- xi + psi*rtruncnorm(n=1, a=0, b=Inf, mean=0, sd=1/sqrt(w[k])) + (sdev/sqrt(w[k]))\%*\%matrix(rnorm(d, mean = 0, sd = 1), nrow=d, ncol=1)
 cat(k, "/", n, " observations simulated\\n", sep="")
}

# Set parameters of G0
hyperG0 <- list()
hyperG0[["b_xi"]] <- rowMeans(z)
hyperG0[["b_psi"]] <- rep(0,d)
hyperG0[["kappa"]] <- 0.001
hyperG0[["D_xi"]] <- 100
hyperG0[["D_psi"]] <- 100
hyperG0[["nu"]] <- d+1
hyperG0[["lambda"]] <- diag(apply(z,MARGIN=1, FUN=var))

 # hyperprior on the Scale parameter of DPM
 a <- 0.0001
 b <- 0.0001

 # do some plots
 doPlot <- FALSE
 nbclust_init <- 30



 ## Data
 ########
 p <- (ggplot(data.frame("X"=z[1,], "Y"=z[2,]), aes(x=X, y=Y))
       + geom_point()
       + ggtitle("Simple example in 2d data")
       +xlab("D1")
       +ylab("D2")
       +theme_bw())
 p


 # Gibbs sampler for Dirichlet Process Mixtures
 ##############################################
 MCMCsample_st <- DPMGibbsSkewT_test1K(z, hyperG0, a, b, N=1000,
 doPlot, nbclust_init, diagVar=FALSE)
 sapply(MCMCsample_st$U_SS_list, "[[", "df")
 sapply(MCMCsample_st$U_SS_list, "[[", "xi")
 sapply(MCMCsample_st$U_SS_list, "[[", "psi")
 plot(density(sapply(MCMCsample_st$U_SS_list, "[[", "df")[500:1000]))
 plot(density(sapply(MCMCsample_st$U_SS_list, "[[", "psi")[2,500:1000]))
}
\author{
Boris Hejblum
}

