% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evalClust_Binder}
\alias{evalClust_Binder}
\title{Evaluate the loss of a point estimate of the partition compared to a gold standard through Binder loss function.}
\usage{
evalClust_Binder(c, gs, a = 1, b = 1)
}
\arguments{
\item{c}{vector of length \code{n} containing the estimated partition
of the  \code{n} observations.}

\item{gs}{vector of length \code{n} containing the  gold standard
partition of the  \code{n} observations.}

\item{a}{penalty for wrong coclustering in \code{c} compared to code{gs}. Defaults is 1.}

\item{b}{penalty for missed coclustering in \code{c} compared to code{gs}. Defaults is 1.}
}
\value{
the cost of the point estimate \code{c} in regard of the
gold standard \code{gs}.
}
\description{
The cost of a point estimate partition is calculated using a pairwise
coincidence loss function.
}
\author{
Boris Hejblum
}
\references{
J.W. Lau & P.J. Green. Bayesian Model-Based Clustering
Procedures, Journal of Computational and Graphical Statistics,
16(3): 526-558, 2007.

D. B. Dahl. Model-Based Clustering for Expression Data via a
Dirichlet Process Mixture Model, in Bayesian Inference for
Gene Expression and Proteomics, K.-A. Do, P. Muller, M. Vannucci
(Eds.), Cambridge University Press, 2006.
}
\seealso{
\link{similarityMat}, \link{clust_est_Binder}
}

