% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flimited.R
\name{Flimited}
\alias{Flimited}
\title{Compute a limited F-measure}
\usage{
Flimited(n_small_clst, pred, ref)
}
\arguments{
\item{n_small_clst}{the size of the}

\item{partition_ref}{}

\item{partition_est}{}
}
\description{
A limited version of F-measure for
}

