% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fmeasure}
\alias{Fmeasure}
\title{Compute F-measure}
\usage{
Fmeasure(pred, ref, select_ref = NULL)
}
\arguments{
\item{pred}{}

\item{ref}{}

\item{select_ref}{}
}
\value{
a list:
 \itemize{
     \item{\code{Fsum}:}{}
     \item{\code{Fmeasures}:}{}
 }
}
\description{
Compute F-measure
}
\author{
Boris P. Hejblum
}
\references{
Aghaeepour, N., Finak, G., Hoos, H., Mosmann,
T. R., Brinkman, R., Gottardo, R., & Scheuermann, R. H.
Critical assessment of automated flow cytometry data analysis techniques.
Nature Methods, 10(3): 228-38, 2013.
}

