% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fmeasure_costC_arma}
\alias{Fmeasure_costC_arma}
\title{C++ implementation of cost computation with Fmeasure as loss function
using the Armadillo library}
\usage{
Fmeasure_costC_arma(c)
}
\arguments{
\item{c}{list of MCMC partitions}
}
\description{
C++ implementation of cost computation with Fmeasure as loss function
using the Armadillo library
}

