% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FmeasureC_no0}
\alias{FmeasureC_no0}
\title{C++ implementation of the F-measure computation without the ref classe 0}
\usage{
FmeasureC_no0(pred, ref)
}
\arguments{
\item{pred}{vector of a predicted partition}

\item{ref}{vector of a reference partition}
}
\description{
Aghaeepour in FlowCAP 1 ignore the reference class labeled "0"
}
\examples{
library(NPflow)
c <- list(c(1,1,2,3,2,3), c(1,1,1,2,3,3),c(2,2,1,1,1,1))
similarityMatC(sapply(c, "["))


\dontrun{
c2 <- list()
for(i in 1:10){
    c2 <- c(c2, list(rmultinom(n=1, size=3000, prob=rexp(n=3000))))
}
c3 <-sapply(c2, "[")
similarityMatC(c3)
}

}
\references{
N Aghaeepour, G Finak, H Hoos, TR Mosmann, RR Brinkman, R Gottardo,
RH Scheuermann, Critical assessment of automated flow cytometry data analysis
techniques, \emph{Nature Methods}, 10(3):228-38, 2013.
}

