% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gibbsDPMslice}
\alias{gibbsDPMslice}
\title{Slice Sampling of the Dirichlet Process Mixture Model}
\usage{
gibbsDPMslice(z, hyperG0, alpha, N, doPlot = TRUE, nbclust_init = 30)
}
\arguments{
\item{z}{data matrix \code{d x n} with \code{d} dimensions in rows
and \code{n} observations in columns.}

\item{hyperG0}{prior mixing distribution.}

\item{alpha}{parameter of the Dirichlet Process.}

\item{N}{number of MCMC iterations.}

\item{doPlot}{logical flag indicating wether to plot MCMC iteration or not.
Default to \code{TRUE}.}

\item{nbclust_init}{number of clusters at initialisation.
Default to 30 (or less if there are less than 30 observations).}
}
\description{
Slice Sampling of the Dirichlet Process Mixture Model
}
\examples{
rm(list=ls())
#Number of data
n <- 100
set.seed(4321)

# Sample data
m <- matrix(nrow=2, ncol=4, c(-1, 1, 0, 2, 1, -2, -1, -2))
p <- c(0.2, 0.1, 0.4, 0.3) # frequence des clusters

library(expm)
s <- array(dim=c(2,2,4))
s[, ,1] <- matrix(nrow=2, ncol=2, c(0.1, 0, 0, 0.1))
s[, ,2] <- matrix(nrow=2, ncol=2, c(0.01, 0, 0, 0.1))
s[, ,3] <- matrix(nrow=2, ncol=2, c(0.1, 0.08, 0.08, 0.1))
s[, ,4] <- .1*diag(2)
c <- rep(0,n)
z <- matrix(0, nrow=2, ncol=n)
for(k in 1:n){
 c[k] = which(rmultinom(n=1, size=1, prob=p)!=0)
 z[,k] <- m[, c[k]] + sqrtm(s[, , c[k]])\%*\%matrix(rnorm(2, mean = 0, sd = 1), nrow=2, ncol=1)
}

 # Set parameters of G0
 hyperG0 <- list()
 hyperG0[["mu"]] <- c(0,0)
 hyperG0[["kappa"]] <- 1
 hyperG0[["nu"]] <- 4
 hyperG0[["lambda"]] <- diag(2)
 # Scale parameter of DPM
 alpha <- 4
 # Number of iterations
 N <- 20
 # do some plots
 doPlot <- TRUE
 nbclust_init <- 30

 # Gibbs sampler for Dirichlet Process Mixtures
 MCMCsample <- gibbsDPMslice(z, hyperG0, alpha, N, doPlot, nbclust_init)
}
\author{
Boris Hejblum
}

