% Generated by roxygen2 (4.0.1): do not edit by hand
\name{invwishrnd}
\alias{invwishrnd}
\title{Sample from a inverse-Wishart distribution}
\usage{
invwishrnd(n, lambda)
}
\arguments{
\item{n}{degrees of freedom}

\item{lambda}{scale parameter}
}
\description{
For internal use only.
}
\keyword{internal}

