% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsNiWlogpdf.R
\name{lgamma_mv}
\alias{lgamma_mv}
\title{Multivariate log gamma function}
\usage{
lgamma_mv(x, p)
}
\arguments{
\item{x}{strictly positive real number}

\item{p}{integer}
}
\description{
Multivariate log gamma function
}

