% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lgamma_mv2C}
\alias{lgamma_mv2C}
\title{C++ implementation of multivariate log gamma function}
\usage{
lgamma_mv2C(x, p)
}
\arguments{
\item{x}{strictly positive real number}

\item{p}{integer}
}
\description{
C++ implementation of multivariate log gamma function
}

