% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lgamma_mvC}
\alias{lgamma_mvC}
\title{C++ implementation of multivariate log gamma function}
\usage{
lgamma_mvC(x, p)
}
\arguments{
\item{x}{strictly positive real number}

\item{p}{integer}
}
\description{
C++ implementation of multivariate log gamma function
}

