% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MAP}
\alias{MAP}
\title{Maximum a Posteriori from MCMC samples of Dirichlet Process Mixture of Skew Normal}
\usage{
MAP(s, from = 1)
}
\arguments{
\item{s}{an DPM object}

\item{from}{an integer giving the first iteration from which MAP is
to be looked for}
}
\description{
Hyperprior on the concentration parameter of the DP
}

