% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summaryDPMMclust}
\alias{plot.summaryDPMMclust}
\alias{print.summaryDPMMclust}
\alias{print.summaryDPMMclust,}
\title{Methods for a summary of a 'DPMMclust' object}
\usage{
\method{print}{summaryDPMMclust}(s, ...)

\method{plot}{summaryDPMMclust}(s, hm = FALSE, nbsim_densities = 50000, ...)
}
\arguments{
\item{s}{a \code{summaryDPMMclust} object}

\item{hm}{logical flag to plot the heatmap of the similarity matrix.
Default is \code{FALSE}.}

\item{nbsim_densities}{the number of simulated observations to be used
to plot the density lines of the clusters in the point estimate partition plot}

\item{...}{further arguments passed to or from other methods}
}
\description{
Methods for a summary of a 'DPMMclust' object
}
\author{
Boris Hejblum
}

