% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postProcess.DPMMclust.R
\name{MLE_N_mmEM}
\alias{MLE_N_mmEM}
\title{Maximum likelihood estimation of mixture of
Normal inverse Wishart distributed observations with an EM algorithm}
\usage{
MLE_N_mmEM(mu_list, S_list, hyperG0, K, maxit = 100, tol = 0.1,
  plot = TRUE)
}
\description{
Maximum likelihood estimation of mixture of
Normal inverse Wishart distributed observations with an EM algorithm
}
\examples{
set.seed(123)
U_mu <- list()
U_Sigma <- list()
U_nu<-list()
U_kappa<-list()

d <- 2
hyperG0 <- list()
hyperG0[["mu"]] <- rep(1,d)
hyperG0[["kappa"]] <- 0.01
hyperG0[["nu"]] <- d+1
hyperG0[["lambda"]] <- diag(d)

for(k in 1:1000){

  NiW <- rNiW(hyperG0, diagVar=FALSE)
  U_mu[[k]] <-NiW[["mu"]]
  U_Sigma[[k]] <-NiW[["S"]]
}


hyperG02 <- list()
hyperG02[["mu"]] <- rep(2,d)
hyperG02[["kappa"]] <- 1
hyperG02[["nu"]] <- d+10
hyperG02[["lambda"]] <- diag(d)/10

for(k in 1001:2000){

  NiW <- rNiW(hyperG02, diagVar=FALSE)
  U_mu[[k]] <-NiW[["mu"]]
  U_Sigma[[k]] <-NiW[["S"]]
}


mlMLE_N_mmEM( U_mu, U_Sigma, hyperG0, K=2)

hyperG0[["mu"]]
hyperG02[["mu"]]
ml$U_mu

hyperG0[["lambda"]]
hyperG02[["lambda"]]
ml$U_lambda

hyperG0[["nu"]]
hyperG02[["nu"]]
ml$U_nu

hyperG0[["kappa"]]
hyperG02[["kappa"]]
ml$U_kappa
}

