% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MLE_skewT_mmEM}
\alias{MLE_skewT_mmEM}
\title{EM MLE for mixture of sNiW}
\usage{
MLE_skewT_mmEM(xi_list, psi_list, S_list, hyperG0, K, maxit = 50, tol = 0.1)
}
\description{
Maximum likelihood estimation of mixture of
Normal inverse Wishart distributed observations with an EM algorithm
}
\examples{
hyperG0 <- list()
hyperG0$b_xi <- c(0.3, -1.5)
hyperG0$b_psi <- c(0, 0)
hyperG0$kappa <- 0.001
hyperG0$D_xi <- 100
hyperG0$D_psi <- 100
hyperG0$nu <- 20
hyperG0$lambda <- diag(c(0.25,0.35))

xi_list <- list()
psi_list <- list()
S_list <- list()
for(k in 1:1000){
 NNiW <- rNNiW(hyperG0, diagVar=FALSE)
 xi_list[[k]] <- NNiW[["xi"]]
 psi_list[[k]] <- NNiW[["psi"]]
 S_list[[k]] <- NNiW[["S"]]
}

hyperG02 <- list()
hyperG02$b_xi <- c(-1, 2)
hyperG02$b_psi <- c(-0.1, 0.5)
hyperG02$kappa <- 0.001
hyperG02$D_xi <- 10
hyperG02$D_psi <- 10
hyperG02$nu <- 3
hyperG02$lambda <- 0.5*diag(2)

for(k in 1001:2000){
 NNiW <- rNNiW(hyperG02, diagVar=FALSE)
 xi_list[[k]] <- NNiW[["xi"]]
 psi_list[[k]] <- NNiW[["psi"]]
 S_list[[k]] <- NNiW[["S"]]
}

mle <- MLE_skewT_mmEM(xi_list, psi_list, S_list, hyperG0, K=2)
mle
}

