% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmNiWpdf.R
\name{mmNiWpdf}
\alias{mmNiWpdf}
\title{multivariate Normal inverse Wishart probability density function for multiple inputs}
\usage{
mmNiWpdf(mu, Sigma, U_mu0, U_kappa0, U_nu0, U_lambda0, Log = TRUE)
}
\arguments{
\item{Log}{logical flag for returning the log of the probability density
function. Defaults is \code{TRUE}.}

\item{x}{data matrix of dimension p x n, p being the dimension of the
data and n the number of data points}

\item{Mu}{mean vectors matrix of dimension p x K, K being the number of
distributions for which the density probability has to be ealuated}

\item{varcovM}{list of length K of variance-covariance matrices,
each of dimensions p x p}

\item{U_Nu0}{vector of length K of degree of freedom parameters}
}
\value{
matrix of densities of dimension K x n
}
\description{
multivariate Normal inverse Wishart probability density function for multiple inputs
}

