% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmNiWpdfC}
\alias{mmNiWpdfC}
\title{C++ implementation of multivariate Normal inverse Wishart probability density function for multiple inputs}
\usage{
mmNiWpdfC(Mu, Sigma, U_Mu0, U_Kappa0, U_Nu0, U_Sigma0, Log = TRUE)
}
\arguments{
\item{Mu}{mean vectors matrix of dimension p x K, K being the number of
distributions for which the density probability has to be ealuated}

\item{U_Nu0}{vector of length K of degree of freedom parameters}

\item{x}{data matrix of dimension p x n, p being the dimension of the
data and n the number of data points}

\item{varcovM}{list of length K of variance-covariance matrices,
each of dimensions p x p}

\item{logical}{flag for returning the log of the probability density
function. Defaults is \code{TRUE}.}
}
\value{
matrix of densities of dimension K x n
}
\description{
C++ implementation of multivariate Normal inverse Wishart probability density function for multiple inputs
}

