% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsNiW.R
\name{mmsNiWlogpdf}
\alias{mmsNiWlogpdf}
\title{structured Normal inverse Wishart (sNiW) probability density function for multiple inputs}
\usage{
mmsNiWlogpdf(U_xi, U_psi, U_Sigma, U_xi0, U_psi0, U_B0, U_Sigma0, U_df0)
}
\arguments{
\item{U_xi}{a list of length n of observed mean vectors, each of dimension p}

\item{U_psi}{a list of length n of observed skew vectors of dimension p}

\item{U_Sigma}{a list of length n of observed covariance matrices, each of dimension p x p}

\item{U_xi0}{a list of length K of mean vector parameters for sNiW, each of dimension p}

\item{U_psi0}{a list of length K of mean vector parameters for sNiW, each of dimension p}

\item{U_B0}{a list of length K of sturcturing matrix parameters for sNiW, each of dimension 2 x 2}

\item{U_Sigma0}{a list of length K of covariance matrix parameters for sNiW, each of dimension p x p}

\item{U_df0}{a list of length K of degrees of freedom parameters for sNiW, each of dimension p x p}
}
\description{
structured Normal inverse Wishart (sNiW) probability density function for multiple inputs
}

