% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mmvsnpdfC_par}
\alias{mmvsnpdfC_par}
\title{Parallel C++ implementation of multivariate skew Normal probability density function for multiple inputs}
\usage{
mmvsnpdfC_par(x, xi, psi, sigma, Log = TRUE, ncores = 1L)
}
\arguments{
\item{x}{data matrix of dimension p x n, p being the dimension of the
data and n the number of data points}

\item{xi}{mean vectors matrix of dimension p x K, K being the number of
distributions for which the density probability has to be ealuated}

\item{psi}{skew parameter vectors matrix of dimension p x K, K being the number of
distributions for which the density probability has to be ealuated}

\item{ncores}{integer number of cores to be used in parallel}

\item{varcovM}{list of length K of variance-covariance matrices,
each of dimensions p x p}

\item{logical}{flag for returning the log of the probability density
function. Defaults is \code{TRUE}.}
}
\value{
matrix of densities of dimension K x n
}
\description{
Parallel C++ implementation of multivariate skew Normal probability density function for multiple inputs
}
\examples{
library(microbenchmark)

K=1000
d=10
z_mat <- NULL
xi_list <- list()
xi_mat <- NULL
psi_list <- list()
psi_mat <- NULL
S_list <- list()

for(i in 1:K){
z_mat <- c(z_mat, c(rep(1.96,d)))
xi_list <- c(xi_list, list(c(rep(-1.5,d))))
xi_mat <- c(xi_mat, c(rep(-1.5,d)))
psi_list <- c(psi_list, list(c(rep(-0.2,d))))
psi_mat <- c(psi_mat, c(rep(-0.2,d)))
S_list <- c(S_list, list(0.33*diag(d)))
}
z_mat <- matrix(z_mat, ncol=K, nrow=d)
xi_mat <- matrix(xi_mat, ncol=K, nrow=d)
psi_mat <- matrix(psi_mat, ncol=K, nrow=d)

microbenchmark(mvsnpdf(x=z_mat, xi=xi_list, psi=psi_list, sigma=S_list, Log=FALSE),
              mmvsnpdfC(x=z_mat, xi=xi_mat, psi=psi_mat, sigma=S_list, Log=FALSE),
              mmvsnpdfC_par(x=z_mat, xi=xi_mat, psi=psi_mat, sigma=S_list, Log=FALSE,
                     ncores = 2),
              times=10L)
microbenchmark(mmvsnpdfC(x=z_mat, xi=xi_mat, psi=psi_mat, sigma=S_list, Log=FALSE),
              mmvsnpdfC_par(x=z_mat, xi=xi_mat, psi=psi_mat, sigma=S_list, Log=FALSE,
                     ncores = 4),
              times=100L)
}

