% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mmvstpdfC}
\alias{mmvstpdfC}
\title{C++ implementation of multivariate Normal probability density function for multiple inputs}
\usage{
mmvstpdfC(x, xi, psi, sigma, df)
}
\arguments{
\item{x}{data matrix of dimension p x n, p being the dimension of the
data and n the number of data points}

\item{xi}{mean vectors matrix of dimension p x K, K being the number of
distributions for which the density probability has to be ealuated}

\item{psi}{skew parameter vectors matrix of dimension p x K, K being the number of
distributions for which the density probability has to be ealuated}

\item{varcovM}{list of length K of variance-covariance matrices,
each of dimensions p x p}

\item{df}{vector of length K of degree of freedom parameters}
}
\value{
matrix of densities of dimension K x n
}
\description{
C++ implementation of multivariate Normal probability density function for multiple inputs
}
\examples{
#skew-normal limit
mmvsnpdfC(x=matrix(rep(1.96,2), nrow=2, ncol=1),
         xi=matrix(c(0, 0)), psi=matrix(c(1, 1),ncol=1), sigma=list(diag(2))
         )
mvstpdf(x=matrix(rep(1.96,2), nrow=2, ncol=1),
       xi=c(0, 0), psi=c(1, 1), sigma=diag(2),
       df=100000000
       )
mmvstpdfC(x=matrix(rep(1.96,2), nrow=2, ncol=1),
         xi=matrix(c(0, 0)), psi=matrix(c(1, 1),ncol=1), sigma=list(diag(2)),
         df=100000000
         )

#non-skewed limit
mmvtpdfC(x=matrix(rep(1.96,2), nrow=2, ncol=1),
        mean=matrix(c(0, 0)), varcovM=list(diag(2)),
        df=10
        )
mmvstpdfC(x=matrix(rep(1.96,2), nrow=2, ncol=1),
         xi=matrix(c(0, 0)), psi=matrix(c(0, 0),ncol=1), sigma=list(diag(2)),
         df=10
         )

microbenchmark(mvstpdf(x=matrix(rep(1.96,2), nrow=2, ncol=1),
                      xi=c(0, 0), psi=c(1, 1),
                      sigma=diag(2), df=10),
              mmvstpdfC(x=matrix(rep(1.96,2), nrow=2, ncol=1),
                        xi=matrix(c(0, 0)), psi=matrix(c(1, 1),ncol=1),
                        sigma=list(diag(2)), df=10),
              times=10000L)
}

