% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmvstpdfC_par}
\alias{mmvstpdfC_par}
\title{ParallelC++ implementation of multivariate Normal probability density function for multiple inputs}
\usage{
mmvstpdfC_par(x, xi, psi, sigma, df, Log = TRUE, ncores = 1L)
}
\arguments{
\item{x}{data matrix of dimension p x n, p being the dimension of the 
data and n the number of data points}

\item{xi}{mean vectors matrix of dimension p x K, K being the number of 
distributions for which the density probability has to be ealuated}

\item{psi}{skew parameter vectors matrix of dimension p x K, K being the number of 
distributions for which the density probability has to be ealuated}

\item{df}{vector of length K of degree of freedom parameters}

\item{varcovM}{list of length K of variance-covariance matrices, 
each of dimensions p x p}

\item{logical}{flag for returning the log of the probability density 
function. Defaults is \code{TRUE}.}
}
\value{
matrix of densities of dimension K x n
}
\description{
ParallelC++ implementation of multivariate Normal probability density function for multiple inputs
}
\examples{

K=1000
d=10
z_mat <- NULL
xi_list <- list()
xi_mat <- NULL
psi_list <- list()
psi_mat <- NULL
S_list <- list()

for(i in 1:K){
z_mat <- c(z_mat, c(rep(1.96,d)))
xi_list <- c(xi_list, list(c(rep(-1.5,d))))
xi_mat <- c(xi_mat, c(rep(-1.5,d)))
psi_list <- c(psi_list, list(c(rep(-0.2,d))))
psi_mat <- c(psi_mat, c(rep(-0.2,d)))
S_list <- c(S_list, list(0.33*diag(d)))
}
z_mat <- matrix(z_mat, ncol=K, nrow=d)
xi_mat <- matrix(xi_mat, ncol=K, nrow=d)
psi_mat <- matrix(psi_mat, ncol=K, nrow=d)

library(microbenchmark)
microbenchmark(mmvstpdfC(x=z_mat, xi=xi_mat, psi=psi_mat, sigma=S_list, df=rep(10,K), Log=FALSE),
              mmvstpdfC_par(x=z_mat, xi=xi_mat, psi=psi_mat, sigma=S_list, df=rep(10,K),  Log=FALSE, 
                     ncores = 4),
              times=50L)

}

