% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mvnpdfC}
\alias{mvnpdfC}
\title{C++ implementation of multivariate Normal probability density function}
\usage{
mvnpdfC(x, mean, varcovM)
}
\arguments{
\item{x}{data matrix}

\item{mean}{mean vector}

\item{varcovM}{variance covariance matrix}
}
\value{
vector of densities
}
\description{
Based on the implementation from Nino Hardt and Dicko Ahmadou
http://gallery.rcpp.org/articles/dmvnorm_arma/
(accessed in August 2014)
}
\examples{
microbenchmark(mvnpdf(x=matrix(1.96), mean=0, varcovM=diag(1)),
              mvnpdfC(x=matrix(1.96), mean=0, varcovM=diag(1)),
              times=10000L)
}

