% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnpdfC_par}
\alias{mvnpdfC_par}
\title{Parallel C++ implementation of multivariate Normal probability density function}
\usage{
mvnpdfC_par(x, mean, varcovM, Log = TRUE, ncores = 1L)
}
\arguments{
\item{x}{data matrix}

\item{mean}{mean vector}

\item{varcovM}{variance covariance matrix}

\item{logical}{flag for returning the log of the probability density 
function. Defaults is \code{TRUE}}
}
\value{
vector of densities
}
\description{
Based on the implementation from Nino Hardt and Dicko Ahmadou
http://gallery.rcpp.org/articles/dmvnorm_arma/ 
(accessed in August 2014)
}
\examples{
mvnpdf(x=matrix(1.96), mean=0, varcovM=diag(1), Log=FALSE)
mvnpdfC(x=matrix(1.96), mean=0, varcovM=diag(1), Log=FALSE)
mvnpdf(x=matrix(1.96), mean=0, varcovM=diag(1))
mvnpdfC(x=matrix(1.96), mean=0, varcovM=diag(1))

library(microbenchmark)
microbenchmark(mvnpdf(x=matrix(1.96), mean=0, varcovM=diag(1), Log=FALSE),
              mvnpdfC(x=matrix(1.96), mean=0, varcovM=diag(1), Log=FALSE),
              times=10000L)     
              
}

