% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mvstpdf}
\alias{mvstpdf}
\title{multivariate skew-t  probability density function}
\usage{
mvstpdf(x, xi, sigma, psi, df)
}
\arguments{
\item{x}{p x n data matrix with n the number of observations and
p the number of dimensions}

\item{xi}{mean vector or list of mean vectors (either a vector,
a matrix or a list)}

\item{sigma}{variance-covariance matrix or list of variance-covariance
matrices (either a matrix or a list)}

\item{psi}{skew parameter vector or list of skew parameter vectors
(either a vector, a matrix or a list)}

\item{df}{a numeric vector or a list of the degrees of freedom
(either a vector or a list)}
}
\description{
multivariate skew-t  probability density function
}
\examples{
mvstpdf(x=matrix(rep(1.96,2), nrow=2, ncol=1),
      xi=c(0, 0), psi=c(1, 1), sigma=diag(2),
      df=100000000
)
mvsnpdf(x=matrix(rep(1.96,2), nrow=2, ncol=1),
      xi=c(0, 0), psi=c(1, 1), sigma=diag(2)
)
}

