% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalinvwishrnd}
\alias{normalinvwishrnd}
\title{Sample from a normal inverse-Wishart distribution
whose parameter are given by the structure hyper}
\usage{
normalinvwishrnd(hyper)
}
\description{
For internal use only.
}
\keyword{internal}

