% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NuMatParC}
\alias{NuMatParC}
\title{C++ implementation of similarity matrix computation using pre-computed distances}
\usage{
NuMatParC(c, d)
}
\arguments{
\item{c}{an MCMC partitions of length \code{n}.}

\item{d}{a symmetric \code{n x n} matrix containing distances
between each group distributions.}
}
\description{
C++ implementation of similarity matrix computation using pre-computed distances
}
\examples{
c <- c(1,1,2,3,2,3)
d <- matrix(runif(length(c)^2),length(c))
NuMatParC(c,d)


}
\author{
Boris Hejblum, Chariff Alkhassim
}
