% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DPM.R
\name{plot_DPM}
\alias{plot_DPM}
\title{Plot of a Dirichlet process mixture of gaussian distribution partition}
\usage{
plot_DPM(z, U_mu = NULL, U_Sigma = NULL, m, c, i, alpha = "?",
  U_SS = NULL, dims2plot = 1:nrow(z), ellipses = ifelse(length(dims2plot)
  < 3, TRUE, FALSE), gg.add = list(theme()))
}
\description{
Plot of a Dirichlet process mixture of gaussian distribution partition
}
\author{
Boris Hejblum
}

