% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DPMsn.R
\name{plot_DPMsn}
\alias{plot_DPMsn}
\title{Plot of a Dirichlet process mixture of skew normal distribution partition}
\usage{
plot_DPMsn(z, c, i = "", alpha = "?", U_SS, dims2plot = 1:nrow(z),
  ellipses = ifelse(length(dims2plot) < 3, TRUE, FALSE),
  gg.add = list(theme()), nbsim_dens = 1000)
}
\arguments{
\item{gg.add}{A list of instructions to add to the ggplot2 instruction.  See \code{\link[ggplot2]{+.gg}}.
Default is \code{list(theme())}, which adds nothing.
to the plot.}
}
\description{
Plot of a Dirichlet process mixture of skew normal distribution partition
}
\author{
Boris Hejblum
}

