% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DPMst.R
\name{plot_DPMst}
\alias{plot_DPMst}
\title{Plot of a Dirichlet process mixture of skew t-distribution partition}
\usage{
plot_DPMst(z, c, i = "", alpha = "?", U_SS, dims2plot = 1:nrow(z),
  ellipses = ifelse(length(dims2plot) < 3, TRUE, FALSE),
  gg.add = list(theme()), nbsim_dens = 1000, nice = FALSE)
}
\arguments{
\item{gg.add}{A list of instructions to add to the ggplot2 instruction.  See \code{\link[ggplot2]{+.gg}}.
Default is \code{list(theme())}, which adds nothing.
to the plot.}

\item{nbsim_dens}{}

\item{nice}{}
}
\description{
Plot of a Dirichlet process mixture of skew t-distribution partition
}
\author{
Boris Hejblum
}

