% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgamma_ss.R
\name{rgamma_ss}
\alias{rgamma_ss}
\title{Sample from a Gamma distribution with a small shape parameter}
\source{
adapted from rgamss.R v2 (04/07/2015) by R. Martin 
https://www4.stat.ncsu.edu/~rmartin/Codes/rgamss.R
}
\usage{
rgamma_ss(n = 1, shape = 1, rate = 1, do.log = TRUE)
}
\arguments{
\item{n}{the number of observations to sample}

\item{shape}{the shape parameter of the gamma distribution. Default is 1.}

\item{rate}{the shape parameter of the gamma distribution. Default is 1.}

\item{do.log}{logical indicating whether the logarithm should be returned 
instead of the sampled value natural scale}
}
\description{
When concentration parameter alpha is close to zero (only 1 non empty 
cluster being sampled), the data augmentation Gibbs sampling from Escobar 
& West relies on the sampling of from a Gamma distribution
}
\details{
For internal use.
}
\examples{
shape <- 0.001
log(rgamma(n = 10, shape = shape))
NPflow:::rgamma_ss(n = 10, shape = shape)

}
\references{
Liu, C., Martin, R., & Syring, N. (2017). Efficient simulation 
from a gamma distribution with small shape parameter. 
Computational Statistics, 32(4), 1767–1775. 
https://doi.org/10.1007/s00180-016-0692-0
}
\keyword{internal}
