% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNiW.R
\name{rNiW}
\alias{rNiW}
\title{Sample from a Normal inverse-Wishart distribution
whose parameter are given by the structure hyper}
\usage{
rNiW(hyper, diagVar)
}
\description{
For internal use only.
}
\keyword{internal}

