% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNNiW.R
\name{rNNiW}
\alias{rNNiW}
\title{Sample from a normal inverse Wishart distribution
whose parameter are given by the structure SufStat}
\usage{
rNNiW(SufStat, diagVar)
}
\description{
For internal use only.
}
\keyword{internal}
