% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rNNiW_SeqPrior}
\alias{rNNiW_SeqPrior}
\title{Sample from a normal inverse Wishart distribution
whose parameter are given by the structure SufStat}
\usage{
rNNiW_SeqPrior(SufStat, diagVar)
}
\description{
For internal use only.
}
\keyword{internal}

