% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleClassC}
\alias{sampleClassC}
\title{C++ implementation of the multinomial sampling from a matrix
of column vectors each containing the sampling probabilities
for their respective draw}
\usage{
sampleClassC(probMat)
}
\arguments{
\item{probMat}{}
}
\description{
C++ implementation of the multinomial sampling from a matrix
of column vectors each containing the sampling probabilities
for their respective draw
}
\keyword{internal}

