% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleClassC_bis_par}
\alias{sampleClassC_bis_par}
\title{Parallel C++ implementation of the multinomial sampling from a matrix
of column vectors each containing the sampling probabilities
for their respective draw}
\usage{
sampleClassC_bis_par(probMat, ncores = 1L)
}
\arguments{
\item{probMat}{}
}
\description{
Parallel C++ implementation of the multinomial sampling from a matrix
of column vectors each containing the sampling probabilities
for their respective draw
}
\details{
Slower than sampleClassC
}
\keyword{internal}

