% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMat.R
\name{similarityMat}
\alias{similarityMat}
\title{Computes the coclustering (or similarity) matrix}
\usage{
similarityMat(c, step = 1)
}
\arguments{
\item{c}{a list of vector of length \code{n}. \code{c[[j]][i]} is 
the cluster allocation of observation \code{i=1...n} at iteration 
\code{j=1...N}.}

\item{step}{provide coclustering every \code{step} iterations.
Default is 1.}
}
\value{
A matrix of size \code{n x n} whose term \code{[i,j]} 
is the proportion of MCMC iterations where observation \code{i} and 
observations \code{j} are allocated to the same cluster.
}
\description{
Computes the coclustering (or similarity) matrix
}
\author{
Boris Hejblum
}

