% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.DPMMclust}
\alias{summary.DPMMclust}
\title{Summarizing Dirichlet Process Mixture Models}
\usage{
\method{summary}{DPMMclust}(x, burnin = 0, gs = NULL, ...)
}
\arguments{
\item{x}{a \code{DPMMclust} object.}

\item{burnin}{the number of MCMC iterations to burn (defaults is half).}

\item{gs}{optionnal vector of length \code{n} containing the gold standard
partition of the \code{n} observations to compare to the point estimate.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{list}:
 \itemize{
     \item{\code{c_est}:}{ a vector of length \code{n}. Point estimate of the partition}
     \item{\code{cost}:}{ a vector of length \code{N}. \code{cost[j]} is the cost
associated to partition \code{c[[j]]}}
     \item{\code{similarity}:}{  matrix of size \code{n x n}. Similarity matrix
(see \link{similarityMat})}
 }
}
\description{
Summarizing Dirichlet Process Mixture Models
}
\author{
Boris Hejblum
}
\seealso{
\link{similarityMat}
}

