% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{traceEpsC}
\alias{traceEpsC}
\title{C++ implementation of residual trace computation step used when sampling the scale}
\usage{
traceEpsC(eps, sigma)
}
\description{
C++ implementation of residual trace computation step used when sampling the scale
}
\keyword{internal}

