% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{traceEpsC_par}
\alias{traceEpsC_par}
\title{Parallel C++ implementation of residual trace computation step used when sampling the scale}
\usage{
traceEpsC_par(eps, sigma, ncores = 1L)
}
\arguments{
\item{eps}{}
}
\description{
Parallel C++ implementation of residual trace computation step used when sampling the scale
}
\keyword{internal}

