% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_SSsn.R
\name{update_SSsn}
\alias{update_SSsn}
\title{Return updated sufficient statistics S of Normal Wishart distribution
with new data matrix z}
\usage{
update_SSsn(z, S, ltn, hyperprior = NULL)
}
\description{
For internal use only.
}
\keyword{internal}

