% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vclust2mcoclustC}
\alias{vclust2mcoclustC}
\title{C++ implementation}
\usage{
vclust2mcoclustC(c)
}
\arguments{
\item{c}{is an MCMC partition}
}
\description{
C++ implementation
}
\examples{
cc <- c(1,1,2,3,2,3)
vclust2mcoclustC(cc)


}
\author{
Chariff Alkhassim
}

