% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishrnd.R
\name{wishrnd}
\alias{wishrnd}
\title{Sample from a Wishart distribution}
\usage{
wishrnd(n, Sigma)
}
\arguments{
\item{n}{degrees of freedom}

\item{Sigma}{scale parameter (matrix)}
}
\description{
For internal use only.
}
\keyword{internal}

