% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn.DPMM.R
\name{burn.DPMMclust}
\alias{burn.DPMMclust}
\title{Burning MCMC iterations from a Dirichlet Process Mixture Model.}
\usage{
burn.DPMMclust(x, burnin = 0, thin = 1)
}
\arguments{
\item{x}{a \code{DPMMclust} object.}

\item{burnin}{the number of MCMC iterations to burn (default is \code{0}).}

\item{thin}{the spacing at which MCMC iterations are kept.
Default is \code{1}, i.e. no thining.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{DPMMclust} object minus the burnt iterations
}
\description{
Utility function for burning MCMC iteration from a \code{DPMMclust} object.
}
\seealso{
\code{\link{summary.DPMMclust}}
}
\author{
Boris Hejblum
}
