% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clusDPM}
\alias{clusDPM}
\title{Unsupervised clustering though Dirichlet Process Mixture models}
\usage{
clusDPM(data, distibution = c("Normal", "skewNormal", "skewStudent"))
}
\arguments{
\item{data}{}

\item{distribution}{}
}
\description{
Wrapping function for mcmc sampling and partition estimation.
}
\seealso{
DPM_GibbsSampler_SkewN

cluster_est_binder
}

