% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusDPM.R
\name{clusDPM}
\alias{clusDPM}
\title{Unsupervised clustering though Dirichlet Process Mixture models}
\usage{
clusDPM(data, distrib = c("gaussian", "skewnorm", "skewt"), ncores = 1)
}
\arguments{
\item{data}{}

\item{distrib}{}

\item{ncores}{}
}
\description{
Wrapping function for mcmc sampling and partition estimation.
}
\seealso{
\code{\link{DPMpost}}

\code{\link{cluster_est_binder}}
}

