% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Fmeasure_costC}
\alias{Fmeasure_costC}
\title{Multiple cost computations with Fmeasure as the loss function}
\usage{
Fmeasure_costC(c)
}
\arguments{
\item{c}{a matrix where each column is one MCMC partition}
}
\value{
a list with the following elements:
\itemize{
 \item{Fmeas}{TODO}
}
}
\description{
C++ implementation of multiple cost computations with Fmeasure as the loss
function using the Armadillo library
}
\examples{
library(NPflow)
c <- list(c(1,1,2,3,2,3), c(1,1,1,2,3,3),c(2,2,1,1,1,1))
#Fmeasure_costC(sapply(c, "["))

\dontrun{
c2 <- list()
for(i in 1:100){
    c2 <- c(c2, list(rmultinom(n=1, size=2000, prob=rexp(n=2000))))
}
Fmeasure_costC(sapply(c2, "["))
}

}

