% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Fmeasure_costC_arma_par}
\alias{Fmeasure_costC_arma_par}
\title{Parallel C++ implementation of cost computation with Fmeasure as loss function
using the Armadillo library}
\usage{
Fmeasure_costC_arma_par(c, ncores = 1L)
}
\arguments{
\item{c}{list of MCMC partitions}
}
\description{
Parallel C++ implementation of cost computation with Fmeasure as loss function
using the Armadillo library
}

