% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FmeasureC_no0}
\alias{FmeasureC_no0}
\title{C++ implementation of the F-measure computation without the reference class 0}
\usage{
FmeasureC_no0(pred, ref)
}
\arguments{
\item{pred}{vector of a predicted partition}

\item{ref}{vector of a reference partition}
}
\description{
Aghaeepour in FlowCAP 1 ignore the reference class labeled "0"
}
\examples{
library(NPflow)
pred <- c(1,1,2,3,2,3)
ref <- c(2,2,0,0,0,3)
FmeasureC(pred, ref)
FmeasureC_no0(pred, ref)

}
\references{
N Aghaeepour, G Finak, H Hoos, TR Mosmann, RR Brinkman, R Gottardo,
RH Scheuermann, Critical assessment of automated flow cytometry data analysis
techniques, \emph{Nature Methods}, 10(3):228-38, 2013.
}
