% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invwishrnd.R
\name{invwishrnd}
\alias{invwishrnd}
\title{Sample from a inverse-Wishart distribution}
\usage{
invwishrnd(n, lambda)
}
\arguments{
\item{n}{degrees of freedom}

\item{lambda}{scale parameter}
}
\description{
For internal use only.
}
\keyword{internal}
