% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DPMMclust.R
\name{print.summaryDPMMclust}
\alias{plot.summaryDPMMclust}
\alias{print.summaryDPMMclust}
\alias{summaryDPMMclust}
\title{Methods for a summary of a 'DPMMclust' object}
\usage{
\method{print}{summaryDPMMclust}(x, ...)

\method{plot}{summaryDPMMclust}(x, hm = FALSE, nbsim_densities = 5000,
  gg.add = list(theme_bw()), ...)
}
\arguments{
\item{x}{a \code{summaryDPMMclust} object.}

\item{...}{further arguments passed to or from other methods}

\item{hm}{logical flag to plot the heatmap of the similarity matrix.
Default is \code{FALSE}.}

\item{nbsim_densities}{the number of simulated observations to be used
to plot the density lines of the clusters in the point estimate partition plot}

\item{gg.add}{a list of instructions to add to the ggplot2 instruction.
See \code{\link[ggplot2]{+.gg}}. Default is \code{list(theme())}, which adds nothing to the plot.}
}
\description{
Methods for a summary of a 'DPMMclust' object
}
\author{
Boris Hejblum
}

