% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DPMMclust.R
\name{print.summaryDPMMclust}
\alias{print.summaryDPMMclust}
\alias{summaryDPMMclust}
\alias{plot.summaryDPMMclust}
\alias{plot.summaryDPMMclust}
\title{Methods for a summary of a 'DPMMclust' object}
\usage{
\method{print}{summaryDPMMclust}(x, ...)

\method{plot}{summaryDPMMclust}(x, hm = FALSE, nbsim_densities = 5000,
  hm_subsample = NULL, hm_order_by_clust = TRUE,
  gg.add = list(theme_bw()), ...)
}
\arguments{
\item{x}{a \code{summaryDPMMclust} object.}

\item{...}{further arguments passed to or from other methods}

\item{hm}{logical flag to plot the heatmap of the similarity matrix.
Default is \code{FALSE}.}

\item{nbsim_densities}{the number of simulated observations to be used
to plot the density lines of the clusters in the point estimate partition plot}

\item{hm_subsample}{a integer designating the number of observations to use when plotting the heatmap. 
Used only if \code{hm} is \code{TRUE}. #'Default is \code{NULL} in which no subsampling is done and 
all observations are plotted.}

\item{hm_order_by_clust}{logical flag indicating whether observations should be ordered according to
the point estimate first. Used only if \code{hm} is \code{TRUE}. Default is \code{TRUE}.}

\item{gg.add}{a list of instructions to add to the ggplot2 instruction (see 
\code{\link[ggplot2]{gg-add}}). Default is \code{list(theme())}, which adds nothing to the plot.}
}
\description{
Methods for a summary of a 'DPMMclust' object
}
\author{
Boris Hejblum
}
