% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mvsnpdf}
\alias{mvsnpdf}
\title{multivariate Skew-Normal  probability density function}
\usage{
mvsnpdf(x, xi, sigma, psi)
}
\arguments{
\item{x}{}

\item{mean}{}

\item{omega}{}
}
\description{
multivariate Skew-Normal  probability density function
}
\examples{
mvsnpdf(x=matrix(1.96), mean=0, omega=diag(1))
dnorm(1.96)

mvsnpdf(x=matrix(rep(1.96,2), nrow=2, ncol=1),
      xi=c(0, 0), psi=c(1, 1), sigma=diag(2)
)
}

