% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mvsnpdf}
\alias{mvsnpdf}
\title{multivariate Skew-Normal probability density function}
\usage{
mvsnpdf(x, xi, sigma, psi)
}
\arguments{
\item{x}{p x n data matrix with n the number of observations and
p the number of dimensions}

\item{xi}{mean vector or list of mean vectors (either a vector,
a matrix or a list)}

\item{sigma}{variance-covariance matrix or list of variance-covariance
matrices (either a matrix or a list)}

\item{psi}{skew parameter vector or list of skew parameter vectors
(either a vector, a matrix or a list)}
}
\description{
multivariate Skew-Normal probability density function
}
\examples{
mvsnpdf(x=matrix(1.96), mean=0, omega=diag(1))
dnorm(1.96)

mvsnpdf(x=matrix(rep(1.96,2), nrow=2, ncol=1),
      xi=c(0, 0), psi=c(1, 1), sigma=diag(2)
)
}

