% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvstlikC_par}
\alias{mvstlikC_par}
\title{Parallel C++ implementation of multivariate Normal probability density function for multiple inputs}
\usage{
mvstlikC_par(x, c, clustval, xi, psi, sigma, df, loglik = TRUE, ncores = 1L)
}
\arguments{
\item{x}{data matrix of dimension p x n, p being the dimension of the
data and n the number of data points}

\item{c}{integer vector of cluster allocations with values from 1 to K}

\item{clustval}{vector of unique values from c in the order corresponding to
the storage of cluster parameters in \code{xi}, \code{psi}, and \code{varcovM}}

\item{xi}{mean vectors matrix of dimension p x K, K being the number of
clusters}

\item{psi}{skew parameter vectors matrix of dimension p x K}

\item{df}{vector of length K of degree of freedom parameters}

\item{varcovM}{list of length K of variance-covariance matrices,
each of dimensions p x p}
}
\value{
a list: "indiv": vector of likelihood of length n; "clust": vector of likelihood of length K; "total": total (log)-likelihood
}
\description{
Parallel C++ implementation of multivariate Normal probability density function for multiple inputs
}
\seealso{
mmvstpdfC, mvstpdf
}

