% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mvstpdf}
\alias{mvstpdf}
\title{multivariate Skew-t  probability density function}
\usage{
mvstpdf(x, xi, sigma, psi, df)
}
\arguments{
\item{x}{}

\item{xi}{}

\item{sigma}{}

\item{psi}{}

\item{df}{}
}
\description{
multivariate Skew-t  probability density function
}

