% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mvtpdf}
\alias{mvtpdf}
\title{multivariate Student's t-distribution probability density function}
\usage{
mvtpdf(x, mean, varcovM, df)
}
\arguments{
\item{x}{data matrix}

\item{mean}{mean vector or list of mean vectors (either a vector,
a matrix or a list)}

\item{varcovM}{}

\item{df}{}
}
\description{
multivariate Student's t-distribution probability density function
}
\examples{
mvnpdf(x=matrix(1.96), mean=0, varcovM=diag(1))
dnorm(1.96)

mvnpdf(x=matrix(rep(1.96,2), nrow=1, ncol=2),
      mean=c(0, 0), varcovM=diag(2)
)
}

