% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NPflow-package.R
\docType{package}
\name{NPflow-package}
\alias{NPflow}
\alias{NPflow-package}
\title{Nonparametrics Bayesian Automatic Gating of Flow Cytometry data}
\description{
This package implements
}
\details{
\tabular{ll}{
Package: \tab NPflow\cr
Type: \tab Package\cr
Version: \tab 0.2.6\cr
Date: \tab 2014-12-27\cr
License:\tab \href{http://www.gnu.org/licenses/lgpl.txt}{LGPL-3}\cr
}
The main function in this package is \code{\link{DPMgibbs}} which
performs TODO
}
\author{
Boris P. Hejblum
--- Maintainer: Boris P. Hejblum
}
\references{
Hejblum, B.P., Caron, F., Thiebaut, R., 2014,
}

