% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPflow-package.R
\docType{package}
\name{NPflow-package}
\alias{NPflow}
\alias{NPflow-package}
\title{Bayesian Nonparametrics for Automatic Gating of Flow Cytometry data}
\description{
This package implements Bayesian estimation for Dirichlet process mixture of skew t-distributions.
}
\details{
\tabular{ll}{
Package: \tab NPflow\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2016-04-02\cr
License:\tab \href{http://www.gnu.org/licenses/lgpl.txt}{LGPL-3}\cr
}
The main functions in this package is \code{\link{DPMpost}}.
}
\author{
Boris P. Hejblum, Chariff Alkhassim, Francois Caron
--- Maintainer: Boris P. Hejblum
}
\references{
BP Hejblum, C Alkhassim, R Gottardo, F Caron, R Thiebaut, Sequential Dirichlet
Process Mixtures of Multivariate Skew t-distributions for Model-based Clustering
of Flow Cytometry Data, submitted 2016.
}

