% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPflow-package.R
\docType{package}
\name{NPflow-package}
\alias{NPflow}
\alias{NPflow-package}
\title{Bayesian Nonparametrics for Automatic Gating of Flow Cytometry data}
\description{
Dirichlet process mixture of multivariate normal, skew-normal or skew t-distributions
modeling oriented towards flow-cytometry data preprocessing applications.
}
\details{
\tabular{ll}{
Package: \tab NPflow\cr
Type: \tab Package\cr
Version: \tab 0.9.0\cr
Date: \tab 2016-04-26\cr
License:\tab \href{http://www.gnu.org/licenses/lgpl.txt}{LGPL-3}\cr
}
The main functions in this package is \code{\link{DPMpost}}.
}
\author{
Boris P. Hejblum, Chariff Alkhassim, Francois Caron
--- Maintainer: Boris P. Hejblum
}
\references{
Hejblum BP, Alkhassim C, Gottardo R, Caron F, Thiebaut R, Sequential Dirichlet
Process Mixtures of Multivariate Skew t-distributions for Model-based Clustering
of Flow Cytometry Data, in preparation.
}

