% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_ConvDPM.R
\name{plot_ConvDPM}
\alias{plot_ConvDPM}
\title{Convergence diagnostic plots}
\usage{
plot_ConvDPM(MCMCsample, from = 1,
  to = length(MCMCsample$logposterior_list), shift = 0, ...)
}
\description{
Convergence diagnostic plots
}

