% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ConvDPM.R
\name{plot_ConvDPM}
\alias{plot_ConvDPM}
\title{Convergence diagnostic plots}
\usage{
plot_ConvDPM(MCMCsample, from = 1,
  to = length(MCMCsample$logposterior_list), shift = 0, thin = 1, ...)
}
\arguments{
\item{MCMCsample}{a \code{DPMMclust} or \code{summaryDPMMclust} object.}

\item{from}{the MCMC iteration from which the plot should start.
Default is \code{1}.}

\item{to}{the MCMC iteration up until which the plot should stop.
Default is \code{1}.}

\item{shift}{a number of initial iterations not to be displayed. Default is \code{0}.}

\item{thin}{integer giving the spacing at which MCMC iterations are kept.
Default is \code{1}, i.e. no thining.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Convergence diagnostic plots
}

