% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleClassC}
\alias{sampleClassC}
\title{C++ implementation of the multinomial sampling from a matrix 
of column vectors each containing the sampling probabilities 
for their respective draw}
\usage{
sampleClassC(probMat)
}
\description{
C++ implementation of the multinomial sampling from a matrix 
of column vectors each containing the sampling probabilities 
for their respective draw
}
\keyword{internal}

