% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sampleClassC_par}
\alias{sampleClassC_par}
\title{Parallel C++ implementation of the multinomial sampling from a matrix
of column vectors each containing the sampling probabilities
for their respective draw}
\usage{
sampleClassC_par(probMat, ncores = 1L)
}
\arguments{
\item{probMat}{}
}
\description{
Parallel C++ implementation of the multinomial sampling from a matrix
of column vectors each containing the sampling probabilities
for their respective draw
}
\keyword{internal}

