% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleClassC_par}
\alias{sampleClassC_par}
\title{Parallel C++ implementation of the multinomial sampling from a matrix 
of column vectors each containing the sampling probabilities 
for their respective draw}
\usage{
sampleClassC_par(probMat, ncores = 1L)
}
\description{
Parallel C++ implementation of the multinomial sampling from a matrix 
of column vectors each containing the sampling probabilities 
for their respective draw
}
\keyword{internal}

