% Generated by roxygen2 (4.0.1): do not edit by hand
\name{traceEpsC}
\alias{traceEpsC}
\title{C++ implementation of residual trace computation step used when sampling the scale}
\usage{
traceEpsC(eps, sigma)
}
\arguments{
\item{eps}{}
}
\description{
C++ implementation of residual trace computation step used when sampling the scale
}
\keyword{internal}

