% Generated by roxygen2 (4.0.2): do not edit by hand
\name{traceEpsC_par}
\alias{traceEpsC_par}
\title{Parallel C++ implementation of residual trace computation step used when sampling the scale}
\usage{
traceEpsC_par(eps, sigma, ncores = 1L)
}
\arguments{
\item{eps}{}
}
\description{
Parallel C++ implementation of residual trace computation step used when sampling the scale
}
\keyword{internal}

