% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_SS.R
\name{update_SS}
\alias{update_SS}
\title{Return updated sufficient statistics S of Normal Wishart distribution
with new data matrix z}
\usage{
update_SS(z, S, hyperprior = NULL)
}
\description{
For internal use only.
}
\keyword{internal}

