% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SS.R
\name{update_SS}
\alias{update_SS}
\title{Return updated sufficient statistics S with new data matrix z}
\usage{
update_SS(z, S, hyperprior = NULL)
}
\arguments{
\item{z}{data matrix}

\item{S}{previous sufficient statistics}

\item{hyperprior}{Default is \code{NULL}}
}
\description{
For internal use only.
}
\keyword{internal}
