% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SSst.R
\name{update_SSst}
\alias{update_SSst}
\title{Return updated sufficient statistics S of Normal Wishart distribution
with new data matrix z}
\usage{
update_SSst(z, S, ltn, scale, df, hyperprior = NULL)
}
\description{
For internal use only.
}
\keyword{internal}

