% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_ssST_SeqPrior.R
\name{update_SSst_SeqPrior}
\alias{update_SSst_SeqPrior}
\title{Return updated sufficient statistics S of Normal Wishart distribution
with new data matrix z}
\usage{
update_SSst_SeqPrior(z, S, ltn, scale, df, hyperprior = NULL)
}
\description{
For internal use only.
}
\keyword{internal}

